Fixed Segfault with Low packets + add Atomics
1. Apparently if we use a burst size that is small (5) then we get
segfaults. The address of the last packet received in a small burst
is garbled, not sure why. This is fixed by simplly changing the burst
to 32 on pktgen and on here.

2. Atomics: Just created atomic struct and wrote to it at the beginning
of each burst. Got help from here: https://github.com/scylladb/dpdk/blob/master/app/test/test_atomic.c

3. rte_pktmbuf_append appends X byte of data to the end of a packet, making
the packet larger it seems. There is an adjust function as well that will
reduce size.

Overall: Confirmed with wireshark that in a burst, packets are incrementing.